+++ packages/themes/gatsby-theme-flexiblog-personal/src/containers/Post.jsx
@@ -12,6 +12,8 @@ import {
   PostImage,
   PostBody,
   PostComments,
+  PostCommentsFacebook,
+  PostCommentsGraph,
   PostTagsShare,
   PostFooter
 } from '@widgets/Post'
@@ -44,6 +46,10 @@ const Post = ({
             <PostBody {...post} />
             <PostTagsShare {...post} location={props.location} />
             {services.disqus && <PostComments {...post} />}
+            {services.graphComment && <PostCommentsGraph {...post} />}
+            {services.facebookComment && (
+              <PostCommentsFacebook {...post} siteUrl={siteUrl} />
+            )}
             <PostFooter {...{ previous, next }} />
           </CardComponent>
         </Main>