+++ starters/gatsby-starter-flexiblog-personal/gatsby-config.js
@@ -6,7 +6,8 @@ module.exports = {
     {
       resolve: '@elegantstack/gatsby-theme-flexiblog-personal',
       options: {
-        // Add theme options here
+        // Add theme options here. Check documentation for available options.
+        siteUrl: process.env.URL || process.env.VERCEL_URL
       }
     }
   ],